refactor: change error message for SSL errors
* rename files to use kebab case
* use object deconstruction where appropriate in the unit tests
