+++ scripts/typedoc/generate_typedoc.sh
@@ -1,6 +1,6 @@
 ./node_modules/.bin/typedoc --mode file --theme ./scripts/typedoc/theme --excludeExternals \
     --out ./doc \
-    ./lib/base_service.ts ./lib/content-type.ts \
+    ./lib/base-service.ts ./lib/content-type.ts \
     ./lib/helper.ts ./lib/querystring.ts \
-    ./lib/requestwrapper.ts \
+    ./lib/request-wrapper.ts \
     ./lib/stream-to-promise.ts --target "ES5"
