+++ test/unit/cp4d-token-manager.test.js
@@ -4,8 +4,8 @@
 const { Cp4dTokenManager } = require('../../auth');
 
 // mock sendRequest
-jest.mock('../../lib/requestwrapper');
-const { RequestWrapper } = require('../../lib/requestwrapper');
+jest.mock('../../lib/request-wrapper');
+const { RequestWrapper } = require('../../lib/request-wrapper');
 const mockSendRequest = jest.fn();
 RequestWrapper.mockImplementation(() => {
   return {
